requirements.txt
asgiref==3.4.1
Django==3.2.6
django-rest-framework==0.1.0
djangorestframework==3.12.4
fasttext==0.9.2
gensim==3.8.3
#gunicorn==20.1.0
compress-fasttext==0.0.7
numpy==1.21.1
pybind11==2.7.1
pytz==2021.1
librabbitmq==2.0.0
sqlparse==0.4.1

amqp==5.0.5
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
django-celery-results==2.0.1
drf-yasg
kombu==5.0.2
librabbitmq==2.0.0
prompt-toolkit==3.0.16
psycopg2==2.8.6
six==1.15.0
SQLAlchemy==1.3.23
vine==5.0.0
wcwidth==0.2.5
